PLUGIN-1515 fix http sink when batch size > 1
Fix the HTTP sink to properly write messages in the OutputFormat
instead of in the transform and destroy methods of the sink,
since the destroy method never gets called in Spark.
This ensures that messages are flushed when the RecordWriter is
closed.

Also fixed it to properly handle unknown input schemas instead of
throwing a NullPointerException
