+++ src/main/java/io/cdap/plugin/http/source/common/http/HttpResponse.java
@@ -80,6 +80,9 @@ public class HttpResponse implements Closeable {
   public byte[] getBytes() {
     if (bytes == null) {
       HttpEntity responseEntity = response.getEntity();
+      if (responseEntity == null) {
+        return new byte[0];
+      }
       try {
         bytes = EntityUtils.toByteArray(responseEntity);
       } catch (IOException e) { // this method is used in multiple next() methods where IOException is not in "throws"
